moved start and stop plugins into timers manager
